feat: add log to file
