+++ src/shared/helpers/file-system.ts
@@ -1,0 +1,7 @@
+import { fileURLToPath } from 'node:url';
+import { dirname } from 'node:path';
+
+export function getCurrentModuleDirectoryPath() {
+  const filepath = fileURLToPath(import.meta.url);
+  return dirname(filepath);
+}
