+++ src/shared/helpers/index.ts
@@ -3,3 +3,5 @@ export {
   getRandomItems,
   getRandomItem
 } from './common.js';
+
+export { getCurrentModuleDirectoryPath } from './file-system.js';
