--- src/shared/libs/logger/pino-logger.ts
@@ -1,28 +1,0 @@
-import { pino, Logger as PinoInstance } from 'pino';
-
-import { Logger } from './logger.interface.js';
-
-
-export class PinoLogger implements Logger {
-  private readonly logger: PinoInstance;
-
-  constructor() {
-    this.logger = pino();
-  }
-
-  public debug(message: string, ...args: unknown[]): void {
-    this.logger.debug(message, ...args);
-  }
-
-  public error(message: string, error: Error, ...args: unknown[]): void {
-    this.logger.error(error, message, ...args);
-  }
-
-  public info(message: string, ...args: unknown[]): void {
-    this.logger.info(message, ...args);
-  }
-
-  public warn(message: string, ...args: unknown[]): void {
-    this.logger.warn(message, ...args);
-  }
-}
